update decomp.bat in accordance to project directory changes + bugfix (#243)
* ObjectFileDB: Fix crash if obj list text has no objects

* rename "decomp.bat" to "decomp-jak1.bat"

+++ decompiler/ObjectFile/ObjectFileDB.cpp
@@ -417,8 +417,10 @@ std::string ObjectFileDB::generate_obj_listing() {
     assert(int(all_unique_names.size()) == unique_count);
   }
 
-  result.pop_back();  // kill last new line
-  result.pop_back();  // kill last comma
+  if (result.length() >= 2) {
+    result.pop_back();  // kill last new line
+    result.pop_back();  // kill last comma
+  }
   return result + "]";
 }
